Add prisma postgres test
